pyproject.toml
[tool.poetry]
name = "steve"
version = "0.1.0"
description = "Cross-continental crypto arbitrage."
authors = [
  "Alex Apt <aapt@uc.cl>",
  "Ariel Martínez <ariel@martinezs.dev>",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
gunicorn = "^20.0.4"
dj-database-url = "^0.5.0"
sentry-sdk = "^1.0.0"
django-cors-headers = "^3.7.0"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
django-extensions = "^3.1.1"
django-admin-numeric-filter = "^0.1.6"
python-binance = "^0.7.11"
trading-api-wrappers = "^0.17.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
pydotplus = "^2.0.2"
isort = "^5.7.0"
ipython = "^7.22.0"

[tool.black]
line-length = 90
target-version = ["py36", "py37", "py38"]
exclude = '/(\.venv|migrations)/'

[build-system]
requires = ["poetry>=1.0.3"]
build-backend = "poetry.masonry.api"
